Figure out permissions
The notebook access-goflex was used to troubleshoot the permissions error. I found that the files themselves required permission to access, so instead of trying to figure that out, I wrote a notebook in the Box where the ECG files were originally stored and then used pandas to write the dataframes to the GoFlex. Data was successfully transferred.