RR interval portion of pipeline
Brought over the pipeline code from the other notebook and modified it to split the output into five files so that the data could be pushed to GitHub.

I copied the code exactly from the other notebook, the only difference being the file splitting, and that code worked well, so I believe we can just go ahead and run the rest! It'll be a few days.